add: product controller, module, schema and service
1) product.schema.ts:
Uses Mongoose ODM (Object Document Mapper) for MongoDB. Defines schema for the "Product" model, which is used to map the structure of data to be stored in MongoDB, including the types of the fields/props.
2) product.service.ts:
ProductService class defines create() method that takes 3 parameters and is responsible for creating a new product document in the database. This service can be used by other components or modules within NestJS app, to perform CRUD operations on the "Product" collection.
3) product.controller.ts:
Provides a route for creating a new product by accepting a POST request and passing the request to the ProductService
4) product.module.ts:
Provides necessary configuration for integrating the ProductController and ProductServices with a MongoDB database using Mongoose.